refactored month to be usable from another module
