+++ src/s1/t3/n1/exercici3/Main.java
@@ -1,0 +1,5 @@
+package s1.t3.n1.exercici3;
+
+public class Main {
+
+}
